+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## [0.2.0-beta2] - Released on 2018-11-10
+
+### Breaking Changes
+
+* Changed interface of `CacheStorePolicy.generateFilename` to make it easier to customize your own cache file structure.
+
+### Others
+
+* Updated document and example.
+* Fixed some bugs.
+
+---
+
 ## [0.1.3-beta] - Released on 2018-11-10
 
 * Some bug fixes.